+++ TrabajoPractico1/funciones.h
@@ -1,9 +1,52 @@
 #include <stdio.h>
-int SumarNumeros(int, int);
-int RestarNumeros(int, int);
-int MultiplicarNumeros(int , int);
-float DividirNumeros(int, int);
+/** \brief Ingreso dos numeros flotantes, los somo y me devuelve un flotante
+ *
+ * \param primerNumero float
+ * \param segundoNumero float
+ * \return float
+ *
+ */
+float SumarNumeros(float, float);
+
+
+/** \brief funcion restar, ingreso dos numeros flotantes, resto y me devuelve un flotante
+ *
+ * \param primerNumero float
+ * \param segundoNumero float
+ * \return float
+ *
+ */
+float RestarNumeros(float, float);
+
+
+/** \brief Funcion multiplicar, ingreso dos flotantes, multiplico, y me devuelve un flotante
+ *
+ * \param primerNumero float
+ * \param segundoNumero float
+ * \return float
+ *
+ */
+float MultiplicarNumeros(float , float);
+
+
+/** \brief funcion dividir donde ingreso dos numeros flotantes, los divido, y me devuelvo un flotante
+ *
+ * \param primerNumero float
+ * \param segundoNumero float
+ * \return float
+ *
+ */
+float DividirNumeros(float, float);
+
+
+/** \brief ingreso un numero, igualo variable de control al numero y voy disminuyendo, multiplico y guardo mi numero factorial
+ *
+ * \param numero int numero que ingrese el usuario
+ * \return long int variable factorial
+ *
+ */
 long int CalcularFactorial(int);
+
